refactor(unsource): organize source internals
