+++ packages/unsource/src/index.ts
@@ -1,22 +1,27 @@
-export { custom } from "./custom.ts"
-export { defineSource, defineSources } from "./define.ts"
-export { file } from "./file.ts"
-export type { FileSourceOptions } from "./file.ts"
-export { github } from "./github.ts"
-export type { GitHubSourceOptions } from "./github.ts"
-export { glob } from "./glob.ts"
-export type { GlobSourceOptions } from "./glob.ts"
-export { markdown } from "./markdown.ts"
+export { defineSource, defineSources } from "./core/define.ts"
 export {
   clearSources,
   getRegisteredSource,
   registerSource,
   registerSources,
   useSource,
-} from "./registry.ts"
-export type * from "./types.ts"
+} from "./core/registry.ts"
+export type * from "./core/types.ts"
 export {
   SourceNotFoundError,
   SourcePathError,
   UnsourceError,
-} from "./errors.ts"
+} from "./core/errors.ts"
+export {
+  custom,
+  file,
+  github,
+  glob,
+  markdown,
+} from "./sources/index.ts"
+export * as source from "./sources/index.ts"
+export type {
+  FileSourceOptions,
+  GitHubSourceOptions,
+  GlobSourceOptions,
+} from "./sources/index.ts"
