+++ packages/unsource/src/sources/custom.ts
@@ -1,4 +1,4 @@
-import type { Source } from "./types.ts"
+import type { Source } from "../core/types.ts"
 
 export function custom<const TSource extends Source>(source: TSource): TSource {
   return source
