+++ packages/unsource/src/sources/github.ts
@@ -1,0 +1,2 @@
+export { github } from "./github/index.ts"
+export type { GitHubSourceOptions } from "./github/types.ts"
