+++ packages/unsource/src/sources/markdown.ts
@@ -1,6 +1,6 @@
 import { file, type FileSourceOptions } from "./file.ts"
 
-import type { Source } from "./types.ts"
+import type { Source } from "../core/types.ts"
 
 export function markdown<const TKey extends string = string>(options: FileSourceOptions<TKey>): Source<TKey> {
   return file<TKey>({
