+++ packages/unsource/test/sources/file-glob-markdown.test.ts
@@ -1,0 +1,55 @@
+import { mkdir, mkdtemp, rm, writeFile } from "node:fs/promises"
+import { tmpdir } from "node:os"
+import { join } from "node:path"
+
+import { afterEach, describe, expect, it } from "vitest"
+
+import {
+  clearSources,
+  defineSources,
+  glob,
+  markdown,
+  registerSources,
+  useSource,
+} from "../../src/index.ts"
+
+const tempDirs: string[] = []
+
+async function createRoot() {
+  const root = await mkdtemp(join(tmpdir(), "vitehub-unsource-root-"))
+  tempDirs.push(root)
+  return root
+}
+
+afterEach(async () => {
+  clearSources()
+  await Promise.all(tempDirs.splice(0).map(path => rm(path, { recursive: true, force: true })))
+})
+
+describe("@vitehub/unsource local file sources", () => {
+  it("loads file, markdown, and glob providers", async () => {
+    const root = await createRoot()
+    await mkdir(join(root, "docs"), { recursive: true })
+    await mkdir(join(root, "node_modules", "pkg"), { recursive: true })
+    await writeFile(join(root, "docs", "README.md"), "# Docs\n")
+    await writeFile(join(root, "docs", "guide.md"), "# Guide\n")
+    await writeFile(join(root, "node_modules", "pkg", "skip.md"), "# Skip\n")
+
+    registerSources(defineSources({
+      docs: glob({ cwd: ".", include: "**/*.md" }),
+      readme: markdown({ path: "docs/README.md", workspacePath: "README.md" }),
+    }))
+
+    await expect(useSource("docs", { rootDir: root }).keys()).resolves.toEqual([
+      "docs/guide.md",
+      "docs/README.md",
+    ])
+    await expect(useSource("readme", { rootDir: root }).get("README.md")).resolves.toMatchObject({
+      mediaType: "text/markdown",
+    })
+    await expect(useSource("readme", { rootDir: root }).read("missing.md" as any))
+      .rejects.toThrow("source.file could not find")
+    await expect(useSource("docs", { rootDir: root }).read("../package.json" as any))
+      .rejects.toThrow("source.glob could not find")
+  })
+})
