+++ packages/unsource/tsdown.config.ts
@@ -3,7 +3,15 @@ import { defineConfig } from "tsdown"
 export default defineConfig({
   clean: true,
   dts: true,
-  entry: ["src/index.ts"],
+  entry: [
+    "src/index.ts",
+    "src/sources/custom.ts",
+    "src/sources/file.ts",
+    "src/sources/github.ts",
+    "src/sources/glob.ts",
+    "src/sources/index.ts",
+    "src/sources/markdown.ts",
+  ],
   exports: {
     inlinedDependencies: false,
   },
